therapist-webapi/requirements.txt
django>=3.2.14
SQLAlchemy===1.4.51
mysqlclient>=2.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.7.0
secure-smtplib>=0.1.1 
pytz>=2024.1
cryptography>=41.0.7
uszipcode===1.0.1